requirements.txt
google-genai>=0.8.0
edge-tts>=6.1.9
moviepy==1.0.3
imageio==2.33.0
imageio-ffmpeg==0.4.9
Pillow>=10.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
pyyaml>=6.0.0
numpy>=1.24.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
